+++ packages/wxt/src/core/utils/env.ts
@@ -2,9 +2,7 @@ import { config } from 'dotenv';
 import { expand } from 'dotenv-expand';
 import type { TargetBrowser } from '../../types';
 
-/**
- * Load environment files based on the current mode and browser.
- */
+/** Load environment files based on the current mode and browser. */
 export function loadEnv(mode: string, browser: TargetBrowser) {
   return expand(
     config({