+++ packages/wxt/src/core/utils/manifest.ts
@@ -7,7 +7,7 @@ import {
   PopupEntrypoint,
   SidepanelEntrypoint,
 } from '../../types';
-import fs from 'fs-extra';
+import { mkdir } from 'node:fs/promises';
 import { resolve } from 'path';
 import { getEntrypointBundlePath } from './entrypoints';
 import { ContentSecurityPolicy } from './content-security-policy';
@@ -23,9 +23,7 @@ import { wxt } from '../wxt';
 import { ManifestV3WebAccessibleResource } from './types';
 import type { Browser } from '@wxt-dev/browser';
 
-/**
- * Writes the manifest to the output directory and the build output.
- */
+/** Writes the manifest to the output directory and the build output. */
 export async function writeManifest(
   manifest: Browser.runtime.Manifest,
   output: BuildOutput,
@@ -35,7 +33,7 @@ export async function writeManifest(
       ? JSON.stringify(manifest)
       : JSON.stringify(manifest, null, 2);
 
-  await fs.ensureDir(wxt.config.outDir);
+  await mkdir(wxt.config.outDir, { recursive: true });
   await writeFileIfDifferent(resolve(wxt.config.outDir, 'manifest.json'), str);
 
   output.publicAssets.unshift({
@@ -44,9 +42,7 @@ export async function writeManifest(
   });
 }
 
-/**
- * Generates the manifest based on the config and entrypoints.
- */
+/** Generates the manifest based on the config and entrypoints. */
 export async function generateManifest(
   allEntrypoints: Entrypoint[],
   buildOutput: Omit<BuildOutput, 'manifest'>,
@@ -155,8 +151,9 @@ export async function generateManifest(
 }
 
 /**
- * Removes suffixes from the version, like X.Y.Z-alpha1 (which browsers don't allow), so it's a
- * simple version number, like X or X.Y or X.Y.Z, which browsers allow.
+ * Removes suffixes from the version, like X.Y.Z-alpha1 (which browsers don't
+ * allow), so it's a simple version number, like X or X.Y or X.Y.Z, which
+ * browsers allow.
  */
 function simplifyVersion(versionName: string): string {
   // Regex adapted from here: https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/manifest.json/version#version_format
@@ -514,8 +511,8 @@ function addDevModePermissions(manifest: Browser.runtime.Manifest) {
 }
 
 /**
- * Returns the bundle paths to CSS files associated with a list of content scripts, or undefined if
- * there is no associated CSS.
+ * Returns the bundle paths to CSS files associated with a list of content
+ * scripts, or undefined if there is no associated CSS.
  */
 export function getContentScriptCssFiles(
   contentScripts: ContentScriptEntrypoint[],
@@ -541,9 +538,9 @@ export function getContentScriptCssFiles(
 }
 
 /**
- * Content scripts configured with `cssInjectionMode: "ui"` need to add their CSS files to web
- * accessible resources so they can be fetched as text and added to shadow roots that the UI is
- * added to.
+ * Content scripts configured with `cssInjectionMode: "ui"` need to add their
+ * CSS files to web accessible resources so they can be fetched as text and
+ * added to shadow roots that the UI is added to.
  */
 export function getContentScriptCssWebAccessibleResources(
   contentScripts: ContentScriptEntrypoint[],
@@ -571,8 +568,8 @@ export function getContentScriptCssWebAccessibleResources(
 }
 
 /**
- * Based on the build output, return a Record of each content script's name to it CSS file if the
- * script includes one.
+ * Based on the build output, return a Record of each content script's name to
+ * it CSS file if the script includes one.
  */
 export function getContentScriptsCssMap(
   buildOutput: Omit<BuildOutput, 'manifest'>,
@@ -610,8 +607,8 @@ function addHostPermission(
 }
 
 /**
- * - "<all_urls>" &rarr; "<all_urls>"
- * - "*://play.google.com/books/*" &rarr; "*://play.google.com/*"
+ * - "<all_urls>" → "<all_urls>"
+ * - "_://play.google.com/books/_" → "_://play.google.com/_"
  */
 export function stripPathFromMatchPattern(pattern: string) {
   const protocolSepIndex = pattern.indexOf('://');
@@ -622,9 +619,10 @@ export function stripPathFromMatchPattern(pattern: string) {
 }
 
 /**
- * Converts all MV3 web accessible resources to their MV2 forms. MV3 web accessible resources are
- * generated in this file, and may be defined by the user in their manifest. In both cases, when
- * targeting MV2, automatically convert their definitions down to the basic MV2 array.
+ * Converts all MV3 web accessible resources to their MV2 forms. MV3 web
+ * accessible resources are generated in this file, and may be defined by the
+ * user in their manifest. In both cases, when targeting MV2, automatically
+ * convert their definitions down to the basic MV2 array.
  */
 export function convertWebAccessibleResourcesToMv2(
   manifest: Browser.runtime.Manifest,
@@ -674,9 +672,7 @@ function convertCspToMv2(manifest: Browser.runtime.Manifest): void {
     manifest.content_security_policy.extension_pages;
 }
 
-/**
- * Make sure all resources are in MV3 format. If not, add a warning.
- */
+/** Make sure all resources are in MV3 format. If not, add a warning. */
 function validateMv3WebAccessibleResources(
   manifest: Browser.runtime.Manifest,
 ): void {
@@ -694,9 +690,7 @@ function validateMv3WebAccessibleResources(
   }
 }
 
-/**
- * Remove keys from the manifest based on the build target.
- */
+/** Remove keys from the manifest based on the build target. */
 function stripKeys(manifest: Browser.runtime.Manifest): void {
   let keysToRemove: string[] = [];
   if (wxt.config.manifestVersion === 2) {