+++ packages/wxt/src/index.ts
@@ -1,5 +1,6 @@
 /**
  * This module contains:
+ *
  * - JS APIs used by the CLI to build extensions or start dev mode.
  * - Helper functions for defining project config.
  * - Types for building and extension or configuring WXT.